requirements.txt
torch
torchvision
timm
natten
einops
numpy
tqdm
pillow
omegaconf
packaging
ftfy
regex
loralib
safetensors
basicsr
opencv-python
pyiqa
torchmetrics
HyperPyYAML
matplotlib
PyYAML